Chances for item usage, Adjusted Configuration
