+++ src/Events.cpp
@@ -225,219 +225,6 @@ namespace Events
 	}
 
 
-    /// <summary>
-    /// Searches the inventory of an Actor for an appropiate potion and applies it.
-	/// This is iterated until the desired actor values are above the threshold
-	/// DEPRECATED
-    /// </summary>
-    /// <param name="_actor">The actor reference to process</param>
-    /// <param name="magicka">Wether magicka should be restored</param>
-    /// <param name="stamina">Wether stamina should be restored</param>
-    /// <param name="health">Wether health should be restored</param>
-    void ActorUsePotion(RE::Actor* _actor, bool magicka, bool stamina, bool health)
-    {
-		auto begin = std::chrono::steady_clock::now();
-		// if there is nothing to do then just return
-		if (!magicka && !stamina && !health)
-			return;
-        // get inventory contents from the reference
-        //std::map<TESBoundObject*, std::pair<Count, std::unique_ptr<InventoryEntryData>>>;
-        auto itemmap = _actor->GetInventory();
-		// get starting point
-		auto iter = itemmap.begin(); 
-		// get endpoint
-		auto end = itemmap.end(); 
-		float mag = _actor->GetActorValue(RE::ActorValue::kMagicka);
-		float stam = _actor->GetActorValue(RE::ActorValue::kStamina);
-		float hea = _actor->GetActorValue(RE::ActorValue::kHealth);
-		RE::AlchemyItem* item = nullptr;
-		RE::EffectSetting* sett = nullptr;
-		int idrunk = 0;
-		// if there are items in the inventory then do something
-		if (iter != end) {
-			LOG_2("{}[UsePotion] trying to find potion");
-			// searches for potions to use
-			while (iter != end && (magicka || stamina || health) && idrunk < Settings::_maxPotionsPerCycle) {
-				//logger::info("idrunk: {}", std::to_string(idrunk));
-				item = iter->first->As<RE::AlchemyItem>();
-				LOG_3("{}[UsePotion] checking item");
-				// get item and check wether it is a potion
-				if (item && !(item->IsPoison()) && !(item->IsFood())) {
-					LOG_3("{}[UsePotion] found medicine");
-					// object is alchemyitem and is a potion
-					// get the actovalue associated with the potion
-					if ((item->avEffectSetting) == nullptr && item->effects.size() == 0) {
-						iter++;
-						continue;
-					}
-					RE::ActorValue avh = RE::ActorValue::kAlchemy;
-					RE::ActorValue avm = RE::ActorValue::kAlchemy;
-					RE::ActorValue avs = RE::ActorValue::kAlchemy;
-					float magh = 0;
-					float magm = 0;
-					float mags = 0;
-					// we will not abort the loop, since the number of effects on one item is normally very
-					// limited, so we don't have much iterations
-					if (item->effects.size() > 0) {
-						for (uint32_t i = 0; i < item->effects.size(); i++) {
-							LOG_4("{}[UsePotion] found array effect");
-							sett = item->effects[i]->baseEffect;
-							// if the primary AV affected by the effect matches one of the following
-							// set the appropiate variables, so we effectively track the restoration of
-							// all relevant stats, for items restoring more than one AV at a time.
-							if (sett && sett->data.primaryAV == RE::ActorValue::kMagicka) {
-								LOG_4("{}[UsePotion] found matching magicka effect");
-								avm = sett->data.primaryAV;
-								magm = item->effects[i]->effectItem.magnitude;
-							} else if (sett && sett->data.primaryAV == RE::ActorValue::kStamina) {
-								LOG_4("{}[UsePotion] found matching stamina effect");
-								avs = sett->data.primaryAV;
-								mags = item->effects[i]->effectItem.magnitude;
-							} else if (sett && sett->data.primaryAV == RE::ActorValue::kHealth) {
-								LOG_4("{}[UsePotion] found matching health effect");
-								avh = sett->data.primaryAV;
-								magh = item->effects[i]->effectItem.magnitude;
-							}
-						}
-					} else {
-						RE::MagicItem::SkillUsageData err;
-						item->GetSkillUsageData(err);
-						switch (item->avEffectSetting->data.primaryAV) {
-						case RE::ActorValue::kHealth:
-							avh = RE::ActorValue::kHealth;
-							magh = err.magnitude;
-							break;
-						case RE::ActorValue::kMagicka:
-							avm = RE::ActorValue::kMagicka;
-							magm = err.magnitude;
-							break;
-						case RE::ActorValue::kStamina:
-							avs = RE::ActorValue::kStamina;
-							mags = err.magnitude;
-							break;
-						}
-					}
-					LOG_3("{}[UsePotion] Drink Potion prepare");
-					RE::ExtraDataList* extra = new RE::ExtraDataList();
-					extra->SetOwner(_actor);
-					bool drunk = false;
-					// based on what kind of potion we have, we will use it
-					if (avh == RE::ActorValue::kHealth) {
-						if (health) {
-							LOG_3("{}[UsePotion] Drink Potion");
-							if(Settings::_CompatibilityMode || Settings::_CompatibilityPotionAnimation)
-							{
-								LOG_3("{}[UsePotion] Compatibility Mode")
-								SKSE::ModCallbackEvent* ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionActorInfo");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = _actor;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-								ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionEvent");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = item;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-							}
-							else
-							{
-								RE::ActorEquipManager::GetSingleton()->EquipObject(_actor, item, extra);
-							}
-							drunk = true;
-							//logger::info("drinking");
-							LOG1_4("{}[UsePotion] value health {}", std::to_string(magh));
-						}
-					}
-					if (avm == RE::ActorValue::kMagicka) {
-						if (magicka) {
-							LOG_3("{}[UsePotion] Drink Potion");
-							if(Settings::_CompatibilityMode || Settings::_CompatibilityPotionAnimation)
-							{
-								LOG_3("{}[UsePotion] Compatibility Mode")
-								SKSE::ModCallbackEvent* ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionActorInfo");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = _actor;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-								ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionEvent");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = item;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-							}
-							else
-							{
-								RE::ActorEquipManager::GetSingleton()->EquipObject(_actor, item, extra);
-							}
-							drunk = true;
-							//logger::info("drinking");
-							LOG1_4("{}[UsePotion] value magicka {}", std::to_string(magm));
-						}
-					}
-					if (avs == RE::ActorValue::kStamina) {
-						if (stamina) {
-							LOG_3("{}[UsePotion] Drink Potion");
-							if (Settings::_CompatibilityMode || Settings::_CompatibilityPotionAnimation) {
-								LOG_3("{}[UsePotion] Compatibility Mode")
-								SKSE::ModCallbackEvent* ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionActorInfo");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = _actor;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-								ev = new SKSE::ModCallbackEvent();
-								ev->eventName = RE::BSFixedString("NPCsDrinkPotionEvent");
-								ev->strArg = RE::BSFixedString("");
-								ev->numArg = 0.0f;
-								ev->sender = item;
-								SKSE::GetModCallbackEventSource()->SendEvent(ev);
-							} else {
-								RE::ActorEquipManager::GetSingleton()->EquipObject(_actor, item, extra);
-							}
-							drunk = true;
-							//logger::info("drinking");
-							LOG1_4("{}[UsePotion] value stamina {}", std::to_string(mags));
-						}
-					}
-					if (drunk) {
-						//logger::info("increasing drunk state");
-						mag += magm;
-						stam += mags;
-						hea += magh;
-						idrunk++;
-					}
-				
-
-					/// TODO build in mechanic to track how much ressuorces have been restored, since AV calculations
-					/// take too long.
-
-
-					// check wether the use of the potin pushed our desired actor values over the threshold
-					// if so disable the appropiate ones, and continue.
-					if (magicka && ACM::GetAVPercentageFromValue(_actor, RE::ActorValue::kMagicka, mag) > Settings::_magickaThresholdUpper) {
-						magicka = false;
-						LOG_3("{}[UsePotion] disabled magicka");
-					}
-					if (health && ACM::GetAVPercentageFromValue(_actor, RE::ActorValue::kHealth, hea) > Settings::_healthThresholdUpper) {
-						health = false;
-						LOG_3("{}[UsePotion] disabled health");
-					}
-					if (stamina && ACM::GetAVPercentageFromValue(_actor, RE::ActorValue::kStamina, stam) > Settings::_staminaThresholdUpper) {
-						stamina = false;
-						LOG_3("{}[UsePotion] disabled stamina");
-					}
-				}
-				// advance to next item if not end of function
-				iter++;
-			}
-		}
-		PROF1_2("{}[PROF] [ActorUsePotion] execution time: {} µs", std::to_string(std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::steady_clock::now() - begin).count()));
-    }   
-
 	/// <summary>
 	/// extracts all poisons from a list of AlchemyItems
 	/// </summary>
@@ -647,6 +434,11 @@ namespace Events
 				}
 				sem_actorreset.release();
 
+				if (Settings::_featUseFood) {
+					// use food at the beginning of the fight to simulate the npc having eaten
+					ACM::ActorUseAnyFood(actor, false);
+				}
+
 				if (actor->IsDead())
 					return EventResult::kContinue;
 
@@ -697,13 +489,13 @@ namespace Events
 	static std::thread* actorhandler = nullptr;
 
 #define UseHealth(t, dur, tolerance) \
-	if (Settings::_featHealthRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kHealth) < Settings::_healthThresholdLower) \
+	if (Settings::_featHealthRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kHealth) < Settings::_healthThreshold) \
 		dur = ACM::ActorUsePotion(std::get<0>(aclist[i]), AlchemyEffect::kHealth);
 #define UseMagicka(t, dur, tolerance) \
-	if (Settings::_featMagickaRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kMagicka) < Settings::_magickaThresholdLower) \
+	if (Settings::_featMagickaRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kMagicka) < Settings::_magickaThreshold) \
 		dur = ACM::ActorUsePotion(std::get<0>(aclist[i]), AlchemyEffect::kMagicka);
 #define UseStamina(t, dur, tolerance) \
-	if (Settings::_featStaminaRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kStamina) < Settings::_staminaThresholdLower) \
+	if (Settings::_featStaminaRestoration && dur < tolerance && ACM::GetAVPercentage(std::get<0>(t), RE::ActorValue::kStamina) < Settings::_staminaThreshold) \
 		dur = ACM::ActorUsePotion(std::get<0>(aclist[i]), AlchemyEffect::kMagicka);
 
 
@@ -763,7 +555,7 @@ namespace Events
 				// handle all registered actors
 				// the list does not change while doing this
 				for (int i = 0; i < aclist.size(); i++) {
-					LOG4_1("{}[CheckActors] [Actor] {}", Utility::GetHex(std::get<0>(aclist[i])->GetFormID()));
+					LOG1_1("{}[CheckActors] [Actor] {}", Utility::GetHex(std::get<0>(aclist[i])->GetFormID()));
 					// if actor is valid and not dead
 					if (std::get<0>(aclist[i]) && !(std::get<0>(aclist[i]))->IsDead()) {
 						// handle potions
@@ -775,24 +567,25 @@ namespace Events
 						durother = std::get<4>(aclist[i]) - Settings::_cycletime;
 						durreg = std::get<5>(aclist[i]) - Settings::_cycletime;
 						// get combined effect for magicka, health, and stamina
-						if (Settings::_featHealthRestoration && durh < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kHealth) < Settings::_healthThresholdLower)
+						if (Settings::_featHealthRestoration && durh < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kHealth) < Settings::_healthThreshold)
 							alch = static_cast<uint64_t>(AlchemyEffect::kHealth);
 						else
 							alch = 0;
-						if (Settings::_featMagickaRestoration && durm < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kMagicka) < Settings::_magickaThresholdLower)
+						if (Settings::_featMagickaRestoration && durm < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kMagicka) < Settings::_magickaThreshold)
 							alch2 = static_cast<uint64_t>(AlchemyEffect::kMagicka);
 						else
 							alch2 = 0;
-						if (Settings::_featStaminaRestoration && durs < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kStamina) < Settings::_staminaThresholdLower)
+						if (Settings::_featStaminaRestoration && durs < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kStamina) < Settings::_staminaThreshold)
 							alch3 = static_cast<uint64_t>(AlchemyEffect::kStamina);
 						else
 							alch3 = 0;
 						// construct combined effect
 						alch |= alch2 | alch3;
 						LOG4_4("{}[CheckActors] check for alchemyeffect {} with current dur health {} dur mag {} dur stam {} ", alch, durh, durm, durs);
+						int counter = 0;
 						// use potions
 						// do the first round
-						if (alch != 0) {
+						if (alch != 0 && (Settings::_UsePotionChance == 100 || rand100(rand) < Settings::_UsePotionChance)) {
 							auto tup = ACM::ActorUsePotion(std::get<0>(aclist[i]), alch);
 							LOG1_2("{}[CheckActors] found potion has Alchemy Effect {}", static_cast<uint64_t>(std::get<1>(tup)));
 							switch (std::get<1>(tup)) {
@@ -812,15 +605,15 @@ namespace Events
 							// do the rest of the rounds
 							for (int c = 1; c < Settings::_maxPotionsPerCycle; c++) {
 								// get combined effect for magicka, health, and stamina
-								if (Settings::_featHealthRestoration && durh < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kHealth) < Settings::_healthThresholdLower)
+								if (Settings::_featHealthRestoration && durh < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kHealth) < Settings::_healthThreshold)
 									alch = static_cast<uint64_t>(AlchemyEffect::kHealth);
 								else
 									alch = 0;
-								if (Settings::_featMagickaRestoration && durm < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kMagicka) < Settings::_magickaThresholdLower)
+								if (Settings::_featMagickaRestoration && durm < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kMagicka) < Settings::_magickaThreshold)
 									alch2 = static_cast<uint64_t>(AlchemyEffect::kMagicka);
 								else
 									alch2 = 0;
-								if (Settings::_featStaminaRestoration && durs < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kStamina) < Settings::_staminaThresholdLower)
+								if (Settings::_featStaminaRestoration && durs < tolerance && ACM::GetAVPercentage(std::get<0>(aclist[i]), RE::ActorValue::kStamina) < Settings::_staminaThreshold)
 									alch3 = static_cast<uint64_t>(AlchemyEffect::kStamina);
 								else
 									alch3 = 0;
@@ -831,32 +624,28 @@ namespace Events
 									switch (std::get<1>(tup)) {
 									case AlchemyEffect::kHealth:
 										durh = std::get<0>(tup) * 1000 > Settings::_MaxDuration ? Settings::_MaxDuration : std::get<0>(tup) * 1000;
+										counter++;
 										break;
 									case AlchemyEffect::kMagicka:
 										durm = std::get<0>(tup) * 1000 > Settings::_MaxDuration ? Settings::_MaxDuration : std::get<0>(tup) * 1000;
+										counter++;
 										break;
 									case AlchemyEffect::kStamina:
 										durs = std::get<0>(tup) * 1000 > Settings::_MaxDuration ? Settings::_MaxDuration : std::get<0>(tup) * 1000;
+										counter++;
 										break;
 									}
-								} else if (durother < tolerance)
-									goto DirJ;
+								}
 								else
 									break;
 							}
-						} else if (durother < tolerance) {
-						DirJ:
-							;
-							// call function to use appropiate potion here
-
-							// we only want to use at max one other potion at a time.
 						}
 
 						uint32_t combatdata = Utility::GetCombatData(std::get<0>(aclist[i]));
 						uint32_t tcombatdata = 0;
 						RE::ActorHandle handle = (std::get<0>(aclist[i]))->currentCombatTarget;
 						
-						if (Settings::_featUsePoisons) {
+						if (Settings::_featUsePoisons && (Settings::_UsePoisonChance == 100 || rand100(rand) < Settings::_UsePoisonChance)) {
 							// handle poisons
 							if (std::get<0>(aclist[i])->IsInFaction(Settings::CurrentFollowerFaction) || std::get<0>(aclist[i]) == RE::PlayerCharacter::GetSingleton()) {
 								if ((combatdata & static_cast<uint32_t>(Utility::CurrentCombatStyle::Mage)) == 0 &&
@@ -944,6 +733,7 @@ namespace Events
 											           static_cast<uint64_t>(Settings::AlchemyEffect::kWeaponSpeedMult) |
 											           static_cast<uint64_t>(Settings::AlchemyEffect::kAttackDamageMult);
 										}
+										LOG1_4("{}[CheckActors] check for poison with effect {}", effects);
 										ACM::ActorUsePoison(std::get<0>(aclist[i]), effects);
 									}
 								}
@@ -953,19 +743,20 @@ namespace Events
 									   (combatdata & static_cast<uint32_t>(Utility::CurrentCombatStyle::HandToHand)) == 0) {
 								// we dont handle a follower, so just let the enemy use any poison they have
 								uint64_t effects = static_cast<uint64_t>(Settings::AlchemyEffect::kAnyPoison);
+								LOG1_4("{}[CheckActors] check for poison with effect {}", effects);
 								ACM::ActorUsePoison(std::get<0>(aclist[i]), effects);
 							}
 							// else Mage or Hand to Hand which cannot use poisons
 						}
 
-						if (Settings::_featUseFortifyPotions) {
-							logger::info("fortify potions stuff");
+						if (Settings::_featUseFortifyPotions && counter < Settings::_maxPotionsPerCycle) {
+							//logger::info("fortify potions stuff");
 
 							if ((std::get<0>(aclist[i])->IsInFaction(Settings::CurrentFollowerFaction) || std::get<0>(aclist[i]) == RE::PlayerCharacter::GetSingleton()) && !(Settings::_EnemyNumberThreshold < aclist.size() || (handle && handle.get() && handle.get().get() && handle.get().get()->GetLevel() >= RE::PlayerCharacter::GetSingleton()->GetLevel() * Settings::_EnemyLevelScalePlayerLevel))) {
 								goto SkipFortify;
 							}
 							// handle fortify potions
-							if (durother < tolerance) {
+							if (durother < tolerance && (Settings::_UseFortifyPotionChance == 100 || rand100(rand) < Settings::_UseFortifyPotionChance)) {
 								// general stuff
 								uint64_t effects = static_cast<uint64_t>(Settings::AlchemyEffect::kHealRate) |
 								                   static_cast<uint64_t>(Settings::AlchemyEffect::kHealRateMult) |
@@ -1069,7 +860,8 @@ namespace Events
 								}
 
 								// std::tuple<int, Settings::AlchemyEffect, std::list<std::tuple<float, int, RE::AlchemyItem*, Settings::AlchemyEffect>>>
-								logger::info("take fortify with effects: {}", Utility::GetHex(effects));
+								//logger::info("take fortify with effects: {}", Utility::GetHex(effects));
+								LOG1_4("{}[CheckActors] check for fortify potion with effect {}", effects);
 								auto tup = ACM::ActorUsePotion(std::get<0>(aclist[i]), effects);
 								switch (std::get<1>(tup)) {
 								case Settings::AlchemyEffect::kHealRate:
@@ -1079,9 +871,11 @@ namespace Events
 								case Settings::AlchemyEffect::kMagickaRateMult:
 								case Settings::AlchemyEffect::kStaminaRateMult:
 									durreg = std::get<0>(tup) * 1000 > Settings::_MaxFortifyDuration ? Settings::_MaxFortifyDuration : std::get<0>(tup) * 1000;
+									LOG1_4("{}[CheckActors] used regeneration potion with tracked duration {}", durreg);
 									break;
 								default:
 									durother = std::get<0>(tup) * 1000 > Settings::_MaxFortifyDuration ? Settings::_MaxFortifyDuration : std::get<0>(tup) * 1000;
+									LOG1_4("{}[CheckActors] used fortify av potion with tracked duration {}", durother);
 									break;
 								}
 							}
